[INTERNAL,TEST] deduplicate XIC/XIM parquet reader helpers
Extract shared Arrow/Parquet reader plumbing into an internal helper used by XICParquetFile and XIMParquetFile.

Keep the public reader and row-struct APIs unchanged, and add regression coverage for LARGE_STRING plus FLOAT/DOUBLE parquet column handling.
