+++ src/openms/source/FORMAT/DATAACCESS/sources.cmake
@@ -17,6 +17,7 @@ set(sources_list
 
 list(APPEND sources_list MSChromatogramParquetConsumer.cpp)
 list(APPEND sources_list MobilogramParquetConsumer.cpp)
+list(APPEND sources_list ParquetReaderHelpers.cpp)
 
 ### add path to the filenames
 set(sources)
